more 64bit stuff
I attempted to massage uintptr_t's instead of working within the constraints of longs and unsigned longs on a 64bit system..  it was hell!  So let's do some backporting and deal with it

+++ WinQuake/d_polyse.c
@@ -129,7 +129,7 @@ void D_PolysetDraw (void)
 						// one extra because of cache line pretouching
 
 	a_spans = (spanpackage_t *)
-			(((uintptr_t)&spans[0] + CACHE_SIZE - 1) & ~(CACHE_SIZE - 1));
+			(((long)&spans[0] + CACHE_SIZE - 1) & ~(CACHE_SIZE - 1));
 
 	if (r_affinetridesc.drawtype)
 	{
